refactor(monitor): move Monitor2D to monitors.py
